move ens_address context manager to web3.utils.ens
