+++ tests/core/contracts/test_contract_init.py
@@ -4,7 +4,7 @@ from web3.exceptions import (
     BadFunctionCallOutput,
     NameNotFound,
 )
-from web3.middleware.names import (
+from web3.utils.ens import (
     contract_ens_addresses,
 )
